Fix ns filepath problems, begin to move functions to new structure
ns lines were not accessing the new structure properly, full directory paths were added.

Existing functions are being moved to more appropriate homes, starting with instructions-printer.

Lein eftest was added to plugins for a more colorful test experience

the last test for is-over has been changed from deftest to def as a way of deactivating the test. The restructure began in the middle of a TDD iteration and the test isn't passing because there's been no attempt to add the functionality to the app that the test is testing for. It will be readded or replaced when the restructure ends and feature-work resumes.