Write player-move function, test*
Testing for this function is incomplete. There should be a test that verifies the loop works with a taken-space in addition to the existing test that tests for good input from the function the first time.

Also, the loop has an error message inside of it, this is not ideal and will be moved later.