Build/test functions for end-game logic
board-full? lives in board and not in rules (as "moves-left" or similar) to consolidate knowledge of board state to the board ns. Also, while the function could be written much shorter (e.g. `(some number? board))`) the current version prevents duplication of knowledge that the empty spaces can be determined by if they're ints or not.