Allow players to choose marker, stop using atoms
Players can now choose their markers and the marker of the computer opponent. The marker must be one character in length and cannot be the same as the other marker.

Player status is now being determined by a "turn" instead of clojure's atom data structure for a more functional approach.