Refactor to pull validation out of input-getter
In order to better follow the SRP and in response to comments on https://github.com/ksabbak/ClojureTTT/pull/17

Smaller changes have also been made in response to those comments.

Display has been changed for the board to go 1-9 instead of 0-8 because it allows for smoother checking and also a lot of people (esp. non-programmers) have a hard time with the idea of zero-first numbering systems.